full-model inference method added to Trainer; moved training steps to separate file
